+++ tests-integration/tests/hll_test/main.rs
@@ -15,5 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
+mod bounds;
 mod union;
 mod update;